requirements.txt
azure-ai-projects>=2.0.0
azure-identity>=1.17.0
azure-storage-blob>=12.20.0
python-dotenv>=1.0.0
fastmcp>=3.0.0
